Scrolling/searching section for terminal + misc.
Add terminal keybinds for next/previous match going up and change
scrollUp/Down keybinds to use 1 line.

Describe copying and pasting in the terminal.

Add Scrolling and Searching subsection for terminal describing the
scrolling keybinds and the process for searching.

Remove oh -p pager from tasklog and syslog in the profile, now that
there is a system for scrolling and searching.

Add workaround for completing directories with spaces.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>