Update module sections, remove prompt theme def
Remove my prompt theme definition from profile and load the PSGallery
module instead.

Import the PackageManagement and PowerShellGet modules early in the
profile, in case they are not automatically imported.

Expand the writing modules and PSGallery sections to cover the new
PackageManagement/PowerShellGet cmdlets and expand the notes on
updating/removing/publishing modules.

Add an alias for the 7zip GUI 7zfm.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>