Remove terminal from choco package list
Remove microsoft-windows-terminal from the list of choco packages to
install as it is now managed by Windows 11.

Leave a commented install command for Windows 10 users.

Also leave a commented command for users who had previously installed
the choco package and then upgraded to Windows 11 to remove the package
from the list without uninstalling it.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>