profile: improve appearance of tasklog/syslog
Pipe output of format-eventlog through out-string, otherwise less
doesn't show linebreaks between entries.

Also stop removing text after newlines, this information can be useful
in the syslog output.

The format is still not ideal in less on my terminal, it may vary for
others, I will try to improve it in the future, but this should be
usable at least.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>