tasks: improve registration code
Use pwsh from PATH instead of embedding the executable path so that the
task does not break when PowerShell Core is updated.

Run ssh directly for the port forwards task instead of wrapping it with
pwsh -command, this is completely unnecessary.

Remove elevation for build task example, this is generally unnecessary.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>