Remove StrawberryPerl, add npm wraps, regen code
Remove StrawberryPerl from the install script as it causes a lot of
problems by overriding build tools. Update the text for it with winget
id and a suggestion for how to deal with the PATH breakage.

Add the path to `npm install -g` installed modules in the profile.

Regenerate the included code in README.md, I forgot to do that last
time.

Set the $vcpkg_toolchain variable to the cmake toolchain file for vcpkg,
more on that later.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>