Improve vim setup section
Move `.vimrc` into an included file and change the Makefile to allow
included files without a `.ps1` extension.

Mention and link the GitHub Copilot plugin.

.vimrc:

Remove setting `ttymouse` in regular vim because that does not work.

Add the vcpkg `x64-windows` dynamic include directory to the path along
with the previously added `-static` include directory.

Add autocommand to set the syntax to `gitcommit.vim` and `textwidth=72`
for Git commits.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>