Describe using busybox as the SHELL for GNU Make
Describe using both Git Bash and busybox as the SHELL for GNU Make to
execute POSIX shell commands and the difference between the two in the
GNU Make section.

Note that single quotes should be used for substitutions returning
paths, and update the repository Makefile to do the same.

Update the repository Makefile to work with busybox and to insert both
make.cmd scripts.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>