install: fix BusyBox shims overwriting programs
In the user install script, install the `busybox-lean` Scoop package
instead of the `busybox` package, which is the same busybox binary but
with no shims.

Create the shims for useful BusyBox utilities that do not come from
other packages, and add a couple of other useful UNIX port packages.

Rewrite the BusyBox section to explain this.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>