Add MSYS2+.bashrc+tmux and busybox .shrc
Add a `.shrc` initialization file for the busybox section.

Add an MSYS2 section with an install script that also installs a sample
`.bashrc` and some packages; and a script to install basic tools for a
build environment.

Add functions for MSYS2 environments to the profile.

Rewrite the tmux section to use the MSYS2 tmux and move the WSL
instructions for tmux into an appendix.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>