Use pwsh script for VS env not bat, `vsenv` func
Fix the accumulation of duplicates in the environment from reloading the
profile because of the use of the Visual Studio batch script to set the
environment for Visual Studio. The profile can now be reloaded as many
times as needed. This is also generally very slightly faster than using
the batch file.

Also add the profile function `vsenv` which takes the architecture as a
parameter and sets up the Visual Studio environment for it, and describe
it.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>