profile: fix setting $env:EDITOR to nvim
Fix regression introduced in 5709bff (profile: remove spaces in path
parts in shortpath, 2025-01-03) which changed setting `$env:EDITOR` to
`nvim` not prefixed by the drive, which does not work in the Msys2
runtime which Git for Windows uses.

Change `realpath` to work the same as `shortpath` but keeping the drive,
and use it to set `$env:EDITOR` to nvim.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>