Update terminal config, rm CLANGARM64 on non-ARM64
Check and update the terminal config and disable the `SHIFT+TAB` and
`CTRL+TAB` default keybindings as well.

Remove a repeated line in the terminal config.

For the MSYS2 build environment install script, do not install
`CLANGARM64` when the host is not ARM64 if the `all` parameter was
passed in.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>